Add teslamate-discovery.sh and README
This change adds the teslamate-discovery.sh script and a README that
explains how to use it.  The script is useful for creating Home
Assistant MQTT Discovery messages for cars that are monitored by
Teslamate.  It configures all of the entities, some additional helpers,
a device tracker, and finally groups them all under a single device.

Signed-off-by: GitHub <noreply@github.com>