Range Type
Previously the range topic was hard coded to always map to the rated
range.  A user reported that there are some models for which rated isn't
a good fit and requested that other range types be available.  This
change adds a '--range-type` flag to the CLI that lets a user select
between estimated, ideal, and rated (default) for the range topic.

[resolves #21]

Signed-off-by: GitHub <noreply@github.com>