Configuration File Locations
Previously, the application would optionally attempt to read YAML
configuration from $XDG_CONFIG_HOME/teslamate-discovery,
$HOME/teslamate-discovery, or ./teslamate-discovery.  Unfortunately,
the binary was named the same thing so putting the binary into your
working directory would result in it being treated as a configuration
file that wasn't YAML and would cause the app to crash.  This wasn't
detected in testing because the binary was never in any of the three
locations that were inspected for configuration.

This change updates the application to look for configuration in a
single location: $XDG_CONFIG_HOME/teslamate-discovery/config.yaml.

[resolves #111]

Signed-off-by: Ben Hale <nebhale@nebhale.com>